Center elements via width or height
CenteredHtml is a little tool to center a given HTML element using
either its width or height and preserving its aspect ratio. If it fills
the window's width without overflow, center the height; if it fills the
window's height without overflow, center the width.

There's probably a pure-CSS way to do this but I couldn't find it as
quickly as I could write this JS module.